application.properties updateed

+++ src/main/java/com/bank/repository/RoleRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.bank.model.Role;
 
-public interface RoleRepo extends CrudRepository<Role, Integer>{
+public interface RoleRepo extends CrudRepository<Role, Long>{
 	
 	Optional<Role> findByName(String name);
